+++ packages/network/src/types.ts
@@ -3,6 +3,7 @@ import { Components, ComponentValue, EntityID, SchemaOf } from "@latticexyz/recs
 import { Cached } from "@latticexyz/utils";
 import { BaseContract, BigNumber, ContractInterface } from "ethers";
 import { Observable } from "rxjs";
+import { SyncState } from "./workers";
 
 export interface NetworkConfig {
   chainId: number;
@@ -12,6 +13,9 @@ export interface NetworkConfig {
   snapshotServiceUrl?: string;
   streamServiceUrl?: string;
   initialBlockNumber?: number;
+  blockExplorer?: string;
+  cacheAgeThreshold?: number;
+  cacheInterval?: number;
 }
 
 export interface ClockConfig {
@@ -120,6 +124,9 @@ export type SyncWorkerConfig = {
   snapshotServiceUrl?: string;
   streamServiceUrl?: string;
   fetchSystemCalls?: boolean;
+  cacheInterval?: number;
+  cacheAgeThreshold?: number;
+  snapshotNumChunks?: number;
 };
 
 export enum ContractSchemaValue {
@@ -250,4 +257,10 @@ export type ContractSchemaValueTypes = {
   [ContractSchemaValue.UINT256_ARRAY]: string[];
   [ContractSchemaValue.BYTES_ARRAY]: string[];
   [ContractSchemaValue.STRING_ARRAY]: string[];
+};
+
+export type SyncStateStruct = {
+  state: SyncState;
+  msg: string;
+  percentage: number;
 };