+++ packages/recs/src/Component.ts
@@ -358,7 +358,7 @@ export function overridableComponent<S extends Schema, M extends Metadata, T = u
       if (prop === "keys") {
         return () => new Set([...target.keys(), ...overriddenEntityValues.keys()]).values();
       }
-      return Reflect.get(target, prop);
+      return Reflect.get(target, prop).bind(target);
     },
   });